Add delete functionality to minifigs for the app
Adds a delete route and middleware function to handle delete
functionality for minifigures. Forms are added to the relevant views to
initiate the delete function.

Also moves a query function to the queries.js file from the
minifigQueries.js file as it was more of a general query than one
related to the minifigs table specifically.