Add 'deleteSet' functionality to the app
Adds a route, a middleware function, and a query to allow the deletion
of a specific set from a theme's selection. This deletion also deletes
any related data to the set from the minifig_inclusions tab.

Also minor fixes to the error handling of some middelware functions in
setController.js