Update minifig name validation and theme choice
Updates minifig name validation so that it allow hyphens and updates the
pattern attribute for the minifig name input form controls to reflect
this. Also removes the form controls that allowed users to select a
theme for the minifigure, now the theme tied to the minifig is the one
from which the 'Add minifigure' form was navigated to.

This change is made in order to avoid the complications of switching the
theme a minifig is tied to while it is simultaneously tied to a set
relating to its original theme.

Refactors middleware functions for the minifigController to remove
unnecessary database queries, specifically relating to these new
changes.