Add errors.ejs as an include statement for forms
Adds the partials directory along with the error.ejs file. The errors
files is added to all relevant forms with an include statement to
display error messages. Also the pattern attribute is removed from
certain text inputs as adding a custom message for these input is not
possible right now anyway.