Add error styling to the updateSetMinifigForm view
Add styling to the errors.ejs partial view in the updateSetMinifigForm
view.

    Also fixed an issue the route, controller, and view had that caused
submitting an update, after one submit had already been rejected,to post
to a nonexistent path. The path route for the post method of the form
was changed to fix this and changes to the appropriate controller
middleware and view were made around the new path. Specifically the
minifigId of the updated minifigure is now taken from a hidden input in
the form.